Move form related test helper functions into their own file
